refactor: move session to independent file
Manage session in a independent file.

Log: move session to independent file
